Fix semantic conflict with changes from master
Also some re-arranging.
